refactor: Consolidate image processing modules and update main function calls
